requirements.txt
flask~=3.1.0
flask_sqlalchemy~=3.1.1
flask_migrate~=4.1.0
flask_jwt_extended~=4.7.1
flask-cors~=5.0.1

Flask-Pydantic~=0.12.0
Pydantic~=2.10.5
email-validator~=2.2.0
pydantic-extra-types~=2.10.2
phonenumbers~=8.13.55

python-dotenv~=1.0.1
pillow~=11.1.0

redis~=5.2.1
psycopg2-binary~=2.9.10